+++ packages/yume-engine/src/pe/index.ts
@@ -10,3 +10,6 @@ export * from './binary-reader.js';
 export * from './import-parser.js';
 export * from './version-parser.js';
 export * from './pe-inspector.js';
+export {
+  PeResourceDecoder,
+} from './resource-decoder.js';